+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [19.0.1] - 2019-03-12
+
+### Added
+- New function `getFileUrl()` will return a URL for reading a particular file
+  from an applications' Gaia bucket.
+
+### Changed
+- Fixed the email authentication scope
+- Fixed multiple bugs with `redirectToSignIn()` and `makeAuthRequest()`
+
 ## [19.0.0] - 2019-02-21
 
 ### Added